+++ BankManagerUIApp/BankManagerUIApp/Supportings/Assets.xcassets/AccentColor.colorset/Contents.json
