+++ BankManagerUIApp/BankManagerUIApp/Supportings/Assets.xcassets/AppIcon.appiconset/Contents.json
