+++ BankManagerUIApp/BankManagerUIApp/Supportings/Assets.xcassets/BMGreen.colorset/Contents.json
