+++ BankManagerUIApp/BankManagerUIApp/Supportings/Assets.xcassets/BMPurple.colorset/Contents.json
