+++ BankManagerUIApp/BankManagerUIApp/Supportings/Assets.xcassets/Contents.json
