+++ BankManagerUIApp/BankManagerUIApp/Supportings/Info.plist
